PrinterGUI
This script after you have edited to include your servers and printer groups will open a GUI that will streamline your printer experience. This is mostly used if you aren't using a print server. In my case this was done due to having 30> terminal servers to add and edit printers to.